Rename OpenXR submodule
